# docs/swagger/swagger.yaml
definitions:
...
  orgs.Membership:
    properties:
      org_id:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  orgs.Organization:
    properties:
      domain:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  rootkeys.RootKey:
    properties:
      api_key:
        type: string
      id:
        type: string
    type: object
  routes.ErrorResponse:
    properties:
      error:
        example: not found
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/orgs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/orgs.Organization'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/orgs.Organization'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/orgs.Organization'
        "400":
          description: missing name
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "409":
          description: organization already exists
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create organization
      tags:
      - organizations
  /v1/orgs/{id}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete organization
      tags:
      - organizations
    get:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orgs.Organization'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get organization
      tags:
      - organizations
  /v1/orgs/{id}/members:
    get:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/orgs.Membership'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List organization members
      tags:
      - organizations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Membership to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/orgs.Membership'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/orgs.Membership'
        "400":
          description: missing user_id or invalid role
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: organization not found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "409":
          description: membership already exists
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Add member to organization
      tags:
      - organizations
  /v1/orgs/{id}/members/{userID}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Remove member from organization
      tags:
      - organizations
  /v1/rootkeys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rootkeys.RootKey'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List root keys
      tags:
      - root-keys
    post:
      consumes:
      - application/json
      parameters:
      - description: Root key to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/rootkeys.RootKey'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rootkeys.RootKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "409":
          description: root key already exists
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create root key
      tags:
      - root-keys
  /v1/rootkeys/{id}:
    delete:
      parameters:
      - description: Root key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete root key
      tags:
      - root-keys
    put:
      consumes:
      - application/json
      parameters:
      - description: Root key ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated root key
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/rootkeys.RootKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rootkeys.RootKey'
        "400":
          description: invalid request or id mismatch
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update root key
      tags:
      - root-keys
...
swagger: "2.0"